경고뜨는거 싫어서 수정함

+++ LaborExperimental/LaborExperimental/ServerMove/serverMain.cpp
@@ -5,8 +5,8 @@
 
 CHAR chMsg[1];
 
-int g_posX = 0;
-int g_posY = 0;
+short g_posX = 0;
+short g_posY = 0;
 
 void move(char input);
